requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
matplotlib>=3.4.0
pyyaml>=5.4.0
psutil>=5.8.0

# ROS2 dependencies
rclpy>=3.0.0
rosbag2_py>=0.15.0

# Machine Learning dependencies (optional)
torch>=1.9.0
scikit-learn>=1.0.0
joblib>=1.0.0

# Development dependencies
pytest>=6.0.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0

# FPGA simulation (optional)
cocotb>=1.6.0
